Split forms into admin and transfer forms
