+++ bagitobjecttransfer/recordtransfer/forms/transfer_forms.py
@@ -1,195 +1,13 @@
+''' Forms specific to transferring files with a new submission '''
 from django import forms
 from django.db.models import Case, When, Value, CharField
 from django.forms import BaseFormSet
-from django.utils import timezone
 from django.utils.translation import gettext
-from django.contrib.auth.forms import UserCreationForm
 
 from django_countries.fields import CountryField
 from django_countries.widgets import CountrySelectWidget
 
-from recordtransfer.models import User, Bag, Right, SourceRole, SourceType, Appraisal, Submission
-
-
-class RecordTransferModelForm(forms.ModelForm):
-    ''' Extends forms.ModelForm to add disabled_fields '''
-
-    disabled_fields = []
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        instance = getattr(self, 'instance', None)
-        if instance and instance.pk and self.disabled_fields:
-            # Disable fields
-            for field in self.disabled_fields:
-                if field in self.fields:
-                    self.fields[field].disabled = True
-
-
-class InlineAppraisalFormSet(forms.models.BaseInlineFormSet):
-    model = Appraisal
-    request = None
-
-    @property
-    def empty_form(self):
-        form = super().empty_form
-        if 'user' in form.fields:
-            form.initial['user'] = self.request.user
-            # form.fields['user'].initial = self.request.user
-        return form
-
-
-class AppraisalForm(RecordTransferModelForm):
-    class Meta:
-        model = Appraisal
-        fields = (
-            'appraisal_type',
-            'statement',
-            'note',
-            'submission',
-            'user',
-        )
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if 'note' in self.fields:
-            self.fields['note'].required = False
-
-        if 'user' in self.fields:
-            self.fields['user'].required = False
-            self.fields['user'].widget.can_add_related = False
-            self.fields['user'].widget.can_change_related = False
-            self.fields['user'].widget.attrs['style'] = 'display:none'
-
-        if hasattr(self, 'instance') and hasattr(self.instance, 'submission') and \
-           'submission' in self.fields:
-            self.fields['submission'].help_text = gettext(
-                '<a href="{0}">Click to view submission</a>'
-            ).format(self.instance.submission.get_admin_change_url())
-
-
-class InlineBagForm(RecordTransferModelForm):
-    ''' Form used to view inline Bag '''
-    class Meta:
-        model = Bag
-        fields = (
-            'bag_name',
-            'bagging_date',
-        )
-
-    title = forms.CharField(required=False)
-
-
-class SubmissionForm(RecordTransferModelForm):
-    class Meta:
-        model = Submission
-        fields = (
-            'submission_date',
-            'bag',
-            'user',
-            'accession_identifier',
-            'review_status',
-            'level_of_detail',
-        )
-
-    disabled_fields = ['submission_date', 'bag', 'user']
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields['accession_identifier'].required = False
-        if hasattr(self, 'instance'):
-            self.fields['bag'].help_text = gettext(
-                '<a href="{0}">Click to view bag</a>'
-            ).format(self.instance.bag.get_admin_change_url())
-        self.fields['bag'].widget.can_add_related = False
-
-
-class BagForm(RecordTransferModelForm):
-    ''' Form used to edit a Bag in the admin '''
-
-    class Meta:
-        model = Bag
-        fields = (
-            'user',
-            'bagging_date',
-            'bag_name',
-            'caais_metadata',
-            'part_of_group',
-            'upload_session',
-        )
-
-    disabled_fields = [
-        'user',
-        'bagging_date',
-        'bag_name',
-        'caais_metadata',
-        'part_of_group',
-        'location',
-        'exists',
-    ]
-
-    title = forms.CharField(required=False)
-    location = forms.CharField(required=False)
-    exists = forms.BooleanField(required=False)
-
-    '''
-    def log_new_event(self, event_type: str, event_note: str = ''):
-        new_event = OrderedDict()
-        new_event['event_type'] = event_type
-        new_event['event_date'] = timezone.now().strftime(r'%Y-%m-%d %H:%M:%S %Z')
-        if self.current_user:
-            new_event['event_agent'] = (f'User {self.current_user.username} '
-                                        f'({self.current_user.email})')
-        else:
-            new_event['event_agent'] = 'Transfer Portal User'
-        new_event['event_note'] = event_note
-        self._bag_metadata['section_5']['event_statement'].append(new_event)
-    '''
-
-
-class SignUpForm(UserCreationForm):
-    ''' Form for a user to create a new account '''
-
-    class Meta:
-        model = User
-        fields = ('username', 'first_name', 'last_name', 'email', 'password1', 'password2')
-
-    def clean(self):
-        ''' Clean data, make sure username and email are not already in use. '''
-        cleaned_data = super().clean()
-        new_username = cleaned_data['username']
-        username_exists = User.objects.filter(username=new_username).first() is not None
-        if username_exists:
-            self.add_error('username', f'The username {new_username} is already in use')
-        new_email = cleaned_data['email']
-        email_exists = User.objects.filter(email=new_email).first() is not None
-        if email_exists:
-            self.add_error('email', f'The email {new_email} is already in use')
-        return cleaned_data
-
-    email = forms.EmailField(max_length=256,
-        required=True,
-        widget=forms.TextInput(),
-        label=gettext('Email'))
-
-    username = forms.CharField(max_length=256,
-        min_length=6,
-        required=True,
-        widget=forms.TextInput(),
-        label=gettext('Username'),
-        help_text=gettext('This is the username you will use to log in to your account'))
-
-    first_name = forms.CharField(max_length=256,
-        min_length=2,
-        required=True,
-        widget=forms.TextInput(),
-        label=gettext('First name'))
-
-    last_name = forms.CharField(max_length=256,
-        min_length=2,
-        required=True,
-        widget=forms.TextInput(),
-        label=gettext('Last name'))
+from recordtransfer.models import SourceRole, SourceType, Right
 
 
 class AcceptLegal(forms.Form):
